Fix_avatar_duplicate_bugs

See merge request cms-intern/sw-project!212

+++ pinterest-server/app/Http/Controllers/Api/UserController.php
@@ -182,7 +182,9 @@ class UserController extends BaseController
 
     public function uploadUserAvatar(Request $request){
         $user = User::where('name', $request['name'])->first();
-        unlink(substr($user->avatar_url, 1));
+        if($user->avatar_url !== "/img/avatar.jpeg"){
+            unlink(substr($user->avatar_url, 1));
+        }
 
         $output_file = "img/". $request["name"] ."Avatar". ((new DateTime())->format('Y-m-d--H:i:s')).".jpeg";
         $ifp = fopen( $output_file, 'wb' );
